+++ src/studio/lib/router.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:qtadmin_studio/models/org.dart';
 import 'package:qtadmin_studio/screens/org_screen.dart';
-import 'package:qtadmin_studio/models/recruitment.dart';
+import 'package:qtadmin_studio/models/human.dart';
 import 'package:qtadmin_studio/screens/recruitment_screen.dart';
 
 class ScreenContext {